requirements.txt
elasticsearch==7.5.1
certifi==2019.11.28
python-dotenv==0.11.0